+++ server/src/main/java/org/allaymc/server/block/component/BlockAnvilBaseComponentImpl.java
@@ -1,25 +1,23 @@
 package org.allaymc.server.block.component;
 
 import org.allaymc.api.block.BlockBehavior;
-import org.allaymc.api.block.BlockPlaceHelper;
-import org.allaymc.api.block.data.BlockFace;
 import org.allaymc.api.block.dto.PlayerInteractInfo;
-import org.allaymc.api.block.property.enums.Attachment;
 import org.allaymc.api.block.type.BlockState;
 import org.allaymc.api.block.type.BlockType;
-import org.allaymc.api.container.impl.GrindstoneContainer;
+import org.allaymc.api.container.impl.AnvilContainer;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.math.position.Position3i;
 import org.allaymc.api.world.Dimension;
+import org.allaymc.api.world.Sound;
 import org.joml.Vector3ic;
 
-import static org.allaymc.api.block.property.type.BlockPropertyTypes.ATTACHMENT;
+import static org.allaymc.api.block.property.type.BlockPropertyTypes.MINECRAFT_CARDINAL_DIRECTION;
 
 /**
  * @author IWareQ
  */
-public class BlockGrindstoneBaseComponentImpl extends BlockBaseComponentImpl {
-    public BlockGrindstoneBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
+public class BlockAnvilBaseComponentImpl extends BlockBaseComponentImpl {
+    public BlockAnvilBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
         super(blockType);
     }
 
@@ -30,17 +28,12 @@ public class BlockGrindstoneBaseComponentImpl extends BlockBaseComponentImpl {
             return true;
         }
 
-        blockState = BlockPlaceHelper.processDirection4Property(blockState, placeBlockPos, placementInfo);
+        dimension.addSound(placeBlockPos, Sound.RANDOM_ANVIL_LAND, 1, 0.8f);
 
-        var clickedFace = placementInfo.blockFace();
-        if (clickedFace == BlockFace.DOWN) {
-            blockState = blockState.setPropertyValue(ATTACHMENT, Attachment.HANGING);
-        } else if (clickedFace == BlockFace.UP) {
-            blockState = blockState.setPropertyValue(ATTACHMENT, Attachment.STANDING);
-        } else {
-            blockState = blockState.setPropertyValue(ATTACHMENT, Attachment.SIDE);
-        }
-
+        blockState = blockState.setPropertyValue(
+                MINECRAFT_CARDINAL_DIRECTION,
+                placementInfo.player().getHorizontalFace().rotateYCCW().toMinecraftCardinalDirection()
+        );
         dimension.setBlockState(placeBlockPos.x(), placeBlockPos.y(), placeBlockPos.z(), blockState, placementInfo);
         return true;
     }
@@ -56,9 +49,9 @@ public class BlockGrindstoneBaseComponentImpl extends BlockBaseComponentImpl {
             return false;
         }
 
-        var grindstoneContainer = new GrindstoneContainer();
-        grindstoneContainer.setBlockPos(new Position3i(interactInfo.clickedBlockPos(), interactInfo.player().getDimension()));
-        grindstoneContainer.addViewer(player);
+        var anvilContainer = new AnvilContainer();
+        anvilContainer.setBlockPos(new Position3i(interactInfo.clickedBlockPos(), interactInfo.player().getDimension()));
+        anvilContainer.addViewer(player);
         return true;
     }
 }