+++ server/src/main/java/org/allaymc/server/container/processor/ConsumeActionProcessor.java
@@ -21,7 +21,7 @@ import static org.allaymc.api.item.type.ItemTypes.AIR;
 @Slf4j
 public class ConsumeActionProcessor implements ContainerActionProcessor<ConsumeAction> {
     @Override
-    public ActionResponse handle(ConsumeAction action, EntityPlayer player, int currentActionIndex, ItemStackRequestAction[] actions, Map<Object, Object> dataPool) {
+    public ActionResponse handle(ConsumeAction action, EntityPlayer player, int currentActionIndex, ItemStackRequestAction[] actions, Map<String, Object> dataPool) {
         // We have validated the recipe in CraftRecipeActionProcessor, so here we can believe the client directly
         var sourceContainer = player.getReachableContainerBySlotType(action.getSource().getContainerName().getContainer());
         var sourceStackNetworkId = action.getSource().getStackNetworkId();