+++ server/src/main/java/org/allaymc/server/container/processor/CraftRecipeActionProcessor.java
@@ -29,7 +29,7 @@ public class CraftRecipeActionProcessor implements ContainerActionProcessor<Craf
     public static final String RECIPE_DATA_KEY = "recipe";
 
     @Override
-    public ActionResponse handle(CraftRecipeAction action, EntityPlayer player, int currentActionIndex, ItemStackRequestAction[] actions, Map<Object, Object> dataPool) {
+    public ActionResponse handle(CraftRecipeAction action, EntityPlayer player, int currentActionIndex, ItemStackRequestAction[] actions, Map<String, Object> dataPool) {
         var recipeNetworkId = action.getRecipeNetworkId();
         if (recipeNetworkId >= EnchantmentOptionGenerator.NETWORK_ID_COUNTER_INITIAL_VALUE) {
             return handleEnchantTableRecipe(player, recipeNetworkId);