+++ server/src/main/java/org/allaymc/server/container/processor/CraftResultDeprecatedActionProcessor.java
@@ -12,7 +12,7 @@ import java.util.Map;
  */
 public class CraftResultDeprecatedActionProcessor implements ContainerActionProcessor<CraftResultsDeprecatedAction> {
     @Override
-    public ActionResponse handle(CraftResultsDeprecatedAction action, EntityPlayer player, int currentActionIndex, ItemStackRequestAction[] actions, Map<Object, Object> dataPool) {
+    public ActionResponse handle(CraftResultsDeprecatedAction action, EntityPlayer player, int currentActionIndex, ItemStackRequestAction[] actions, Map<String, Object> dataPool) {
         return null;
     }